+++ NEWS.md
@@ -11,6 +11,7 @@ private alleles are desired instead.
 ### Minor
 * Updated the warning messages returned when importing with potentially problematic metadata to be more descriptive.
 * Checking facet requests for hybrid facets (like `pop.fam` or `pop.chr`) will now double check for any duplicates and error if detected. This is a bit slower but will be safer.
+* Added the `facet.order` argument to `plot_pairwise_fst_heatmap()` function to control the order of subfacet plotting.
 
 ## Bug Fixes
 * Fixed `format_snps()` to avoid any scientific notation. Added `contig` flags to the headers printed with `format = "vcf"`.